enable-source: move to own source files
