+++ suricata/update/sources.py
@@ -134,86 +134,6 @@ def get_enabled_sources():
 
     return sources
 
-def load_sources(config):
-    sources_cache_filename = os.path.join(
-        config.get_cache_dir(), SOURCE_INDEX_FILENAME)
-    if os.path.exists(sources_cache_filename):
-        index = yaml.load(open(sources_cache_filename).read())
-        return index["sources"]
-    return {}
-
-def enable_source(config):
-    name = config.args.name
-
-    # Check if source is already enabled.
-    enabled_source_filename = os.path.join(
-        get_source_directory(), "%s.yaml" % (safe_filename(name)))
-    if os.path.exists(enabled_source_filename):
-        logger.error("The source %s is already enabled.", name)
-        return 1
-
-    # First check if this source was previous disabled and then just
-    # re-enable it.
-    disabled_source_filename = os.path.join(
-        get_source_directory(), "%s.yaml.disabled" % (safe_filename(name)))
-    if os.path.exists(disabled_source_filename):
-        logger.info("Re-enabling previous disabled source for %s.", name)
-        os.rename(disabled_source_filename, enabled_source_filename)
-        return 0
-
-    if not os.path.exists(get_index_filename(config)):
-        logger.warning(
-            "Source index does not exist, "
-            "try running suricata-update update-sources.")
-        return 1
-
-    sources = load_sources(config)
-    if not config.args.name in sources:
-        logger.error("Unknown source: %s", config.args.name)
-        return 1
-
-    # Parse key=val options.
-    opts = {}
-    for param in config.args.params:
-        key, val = param.split("=", 1)
-        opts[key] = val
-
-    source = sources[config.args.name]
-
-    if "subscribe-url" in source:
-        print("The source %s requires a subscription. Subscribe here:" % (name))
-        print("  %s" % source["subscribe-url"])
-
-    params = {}
-    if "parameters" in source:
-        for param in source["parameters"]:
-            if param in opts:
-                params[param] = opts[param]
-            else:
-                prompt = source["parameters"][param]["prompt"]
-                while True:
-                    r = raw_input("%s (%s): " % (prompt, param))
-                    r = r.strip()
-                    if r:
-                        break
-                params[param] = r.strip()
-    new_source = SourceConfiguration(name, params=params).dict()
-
-    if not os.path.exists(get_source_directory()):
-        try:
-            logger.info("Creating directory %s", get_source_directory())
-            os.makedirs(get_source_directory())
-        except Exception as err:
-            logger.error("Failed to create directory %s: %s",
-                         get_source_directory(), err)
-            return 1
-
-    filename = os.path.join(
-        get_source_directory(), "%s.yaml" % (safe_filename(name)))
-    logger.info("Writing %s", filename)
-    with open(filename, "w") as fileobj:
-        fileobj.write(yaml.dump(new_source, default_flow_style=False))
-
 def disable_source(config):
     name = config.args.name
     filename = os.path.join(get_source_directory(), "%s.yaml" % (
